Change a long long to long in PRPACK
R CMD check is complaining that long long is
not ISO C++ 1998.

+++ src/prpack/prpack_result.h
@@ -13,7 +13,7 @@ namespace prpack {
             double read_time;
             double preprocess_time;
             double compute_time;
-            long long num_es_touched;
+            long num_es_touched;
             const char* method;
             int converged;
             // constructor
